Rename module words -> words.local
Golang modules require "." in their name. It is not an issue with a
single main package. However to use subpackages for libraries (for
example for telegram interactions), we need to include "." in the name.

For example, possible package after this change can be
"words.local/telegram". I am reluctant to use
github.com/Uran198/words-telegram-bot as a module name because it is
very long, and I think codebase should not generally know or care where
it's source code is uploaded to.